[FEATURE] Implement post skeletons
