+++ src/modules/posts/post/actions/deletePost.ts
@@ -2,9 +2,14 @@
 
 import { revalidatePath } from 'next/cache';
 
+import { auth } from '@/auth';
 import { db } from '@/lib/database';
 
 export const deletePost = async (postId: string, userId: string) => {
+  const session = await auth();
+
+  if (!session) throw new Error('unauthorized');
+
   const post = await db.post.findUnique({
     where: {
       id: postId,
