+++ src/modules/posts/post/actions/index.ts
@@ -1,5 +1,4 @@
 export { commentPost } from './commentPost';
 export { deletePost } from './deletePost';
-export { fetchPosts } from './fetchPosts';
 export { likePost } from './likePost';
 export { savePost } from './savePost';
