+++ src/modules/posts/post/components/Posts.tsx
@@ -1,8 +1,14 @@
-import { Post } from './post/Post';
-import { fetchPosts } from '../api';
+import { Post } from './post';
+import { fetchPosts } from '../data';
 
 export const Posts = async () => {
   const posts = await fetchPosts();
 
-  return posts.map((post) => <Post key={post.id} {...post} />);
+  return (
+    <>
+      {posts.map((post) => (
+        <Post key={post.id} {...post} />
+      ))}
+    </>
+  );
 };
