+++ src/modules/posts/post/components/PostsSkeleton.tsx
@@ -1,0 +1,5 @@
+import { PostSkeleton } from './post';
+
+export const PostsSkeleton = () => {
+  return [...Array(5)].map((_, index) => <PostSkeleton key={index} />);
+};
