+++ src/modules/posts/post/components/index.ts
@@ -1,1 +1,2 @@
 export { Posts } from './Posts';
+export { PostsSkeleton } from './PostsSkeleton';
