+++ src/modules/posts/post/components/post/comments/bottom-sheet/footer/PostCommentsBottomSheetFooter.tsx
@@ -5,7 +5,7 @@ import showToast from 'react-hot-toast';
 
 import { PaperPlaneIcon } from '@/assets/ui/icons';
 import { useCurrentUser } from '@/hooks';
-import { commentPost } from '@/modules/posts/post/api';
+import { commentPost } from '@/modules/posts/post/actions';
 
 interface PostCommentsBottomSheetFooterProps {
   addOptimisticComment: (action: unknown) => void;
