+++ src/modules/posts/post/components/post/index.ts
@@ -1,0 +1,2 @@
+export { Post } from './Post';
+export { PostSkeleton } from './skeleton/PostSkeleton';
