+++ src/modules/posts/post/data/fetchPosts.ts
@@ -1,5 +1,3 @@
-'use server';
-
 import { revalidatePath, unstable_noStore } from 'next/cache';
 
 import { db } from '@/lib/database';
