+++ src/modules/posts/post/data/index.ts
@@ -1,0 +1,1 @@
+export { fetchPosts } from './fetchPosts';
